* Fixed printf() warnings in src/utility.c * The Makefile install target now no longer strips binaries * Got rid of the CFLAGS_EXTRA variable in the Makefile * Updated INSTALL and README to reflect new Makefile usage
 -- Simrun Basuita  Mon, 15 Dec 2008 17:17:42 +0000

+++ trunk/src/utility.c
@@ -25,8 +25,8 @@ void *malloc_good(size_t size)
   r = malloc(size);
   if (r == NULL)
   {
-    printf("\n\nmalloc_good: Unable to malloc(%i), auto-exiting\n\n",
-                                                                  size);
+    printf("\n\nmalloc_good: Unable to malloc(%lu), auto-exiting\n\n",
+                                                  (unsigned long) size);
     exit(1);
   }
 
@@ -40,8 +40,8 @@ void *realloc_good(void *ptr, size_t size)
   r = realloc(ptr, size);
   if (size != 0 && r == NULL)
   {
-    printf("\n\nrealloc_good: Unable to realloc(%i), auto-exiting\n\n",
-                                                                  size);
+    printf("\n\nrealloc_good: Unable to realloc(%lu), auto-exiting\n\n",
+                                                     (unsigned long) size);
     exit(1);
   }
